feat: rating bar

+++ app/src/main/java/com/company/app/ui/services/Service.kt
@@ -3,7 +3,6 @@ package com.company.app.ui.services
 import android.graphics.drawable.Drawable
 
 data class Service(
-        val score: Double,
+        val score: Float,
         val name : String,
-        val background: Int
-)
+        val background: Int)
